abci/app/abci_app.py
#!/usr/bin/env python3
# MythOS ABCI App for Consensus Cell
# Uses CometBFT (Tendermint successor) for BFT/state machines.

from cometbft.abci.application import Application
from cometbft.abci.types import ResponseInitChain, ResponseQuery

class MythOSApp(Application):
    def __init__(self):
        super().__init__()
        self.state = {}  # Capabilities and manifests

    def init_chain(self, request):
        return ResponseInitChain()

    def query(self, request):
        # Stub: Query for cap manifest
        key = request.data.decode('utf-8')
        value = self.state.get(key, b'not found')
        return ResponseQuery(value=value)

if __name__ == "__main__":
    print("MythOS ABCI App. To run, use a dedicated ABCI server, for example:\n$ uvicorn abci_app:app --host 0.0.0.0 --port 26658")
    # app = MythOSApp()
    # In a real scenario, this would be run by a server like uvicorn or gunicorn.
    # For example: uvicorn abci_app:app --host 0.0.0.0 --port 26658
    # The 'app' object needs to be exposed for the server to find it.
    # Let's assign it for clarity, though running this script directly won't start the server.
    app = MythOSApp()
